@@ -4,11 +4,13 @@ Many on-the-fly transformations which are used to manipulate trajectories as
 they are read.  This allows a trajectory to avoid periodic-boundary issues
 and to automatically align the system to a protein structure.
 """
-import numpy as np
-from numpy import typing as npt
+
 import MDAnalysis as mda
-from MDAnalysis.transformations.base import TransformationBase
+import numpy as np
 from MDAnalysis.analysis.align import rotation_matrix
+from MDAnalysis.lib import distances
+from MDAnalysis.transformations.base import TransformationBase
+from numpy import typing as npt
 
 
 class NoJump(TransformationBase):
@@ -18,9 +20,9 @@ class NoJump(TransformationBase):
     border between two subsequent frames.  This then simplifies the calculation
     of motion over time.
     """
+
     ag: mda.AtomGroup
     prev: npt.NDArray
-
 
     def __init__(self, ag: mda.AtomGroup):
         super().__init__()
@@ -43,30 +45,31 @@ class NoJump(TransformationBase):
         return ts
 
 
-class Minimiser(TransformationBase):
-    """Minimises the difference from ags to central_ag by choosing image
-
-    This transformation will translate any AtomGroup in *ags* in multiples of
-    the box vectors in order to minimise the distance between the center of mass
-    to the center of mass of each ag.
+class ClosestImageShift(TransformationBase):
     """
-    central_ag: mda.AtomGroup
-    other_ags: list[mda.AtomGroup]
+    PBC-safe transformation that shifts one or more target AtomGroups
+    so that their COM is in the closest image relative to a reference AtomGroup.
+    Works for any box type (triclinic or orthorhombic).
 
-    def __init__(self, central_ag: mda.AtomGroup, *ags):
+    CAVEAT:
+    This Transformation requires the AtomGroups to be unwrapped!
+
+    Inspired from:
+    https://github.com/wolberlab/OpenMMDL/blob/main/openmmdl/openmmdl_simulation/scripts/post_md_conversions.py
+    """
+
+    def __init__(self, reference: mda.AtomGroup, targets: list[mda.AtomGroup]):
         super().__init__()
-        self.central_ag = central_ag
-        self.other_ags = ags
+        self.reference = reference
+        self.targets = targets
 
     def _transform(self, ts):
-        center = self.central_ag.center_of_mass()
-        box = self.central_ag.dimensions[:3]
+        center = self.reference.center_of_mass()
 
-        for ag in self.other_ags:
+        for ag in self.targets:
             vec = ag.center_of_mass() - center
-
-            # this only works for orthogonal boxes
-            ag.positions -= np.rint(vec / box) * box
+            vec_min = distances.minimize_vectors(vec.reshape(1, 3), ts.dimensions)[0]
+            ag.translate(vec_min - vec)
 
         return ts
 
@@ -77,6 +80,7 @@ class Aligner(TransformationBase):
     centers all coordinates onto origin
     rotates **entire universe** to minimise rmsd relative to **ref_ag**
     """
+
     ref_pos: npt.NDArray
     ref_idx: npt.NDArray
     weights: npt.NDArray
@@ -98,8 +102,7 @@ class Aligner(TransformationBase):
         mobile_pos -= mobile_com
 
         # rotates mobile to best align with ref
-        R, min_rmsd = rotation_matrix(mobile_pos, self.ref_pos,
-                                      weights=self.weights)
+        R, min_rmsd = rotation_matrix(mobile_pos, self.ref_pos, weights=self.weights)
 
         # apply the transformation onto **all** atoms
         ts.positions -= mobile_com
